refactor(runtime): extract tool_runner helpers (taint, shell_safety, system)
Move tool_runner.rs into tool_runner/mod.rs as the first step of #3710's
god-crate split. Carve out three pure-leaf helper modules:

- `taint`: outbound TaintSink checks (shell_exec, net_fetch, outbound_text,
  outbound_header) plus their secret-key/secret-header denylists.
- `shell_safety`: RO-workspace classifier for shell_exec (quote-aware
  tokenizer, RoSafety enum, classify_shell_exec_ro_safety + segment).
- `system`: location_get and system_time leaf tools.

Behavior-preserving: only `pub(super)` visibility was added so the parent
mod.rs can keep its existing call sites unchanged via `use self::*::*`.
No call path, no public surface, and no test was touched.

mod.rs shrinks from 14,337 LOC to 13,451 LOC (-886). Verification:
cargo check -p librefang-runtime and
cargo clippy -p librefang-runtime --lib -- -D warnings both clean.

Pre-existing failures in tests/mcp_oauth_integration.rs and
tests/tool_runner_forwarding.rs (missing mcp_oauth module and
OAuthTokens type) reproduce on origin/main HEAD and are out of scope
for this refactor.

Refs #3710
